fix: Invalid cluster credentials error is logged with ERROR.
